Разбивает большие вьюхи на несколько маленьких, добавляет обработчики
