--- src/view/film-details-view.js
@@ -1,238 +1,0 @@
-import AbstractView from './abstract-view';
-
-import {
-  getTimeFromMinutes,
-  getFormattedDate,
-  getFormattedDateTime
-} from 'utils/film-card-util';
-
-const getComments = (comments) => (id) => {
-  const {
-    date,
-    author,
-    text,
-    emotion,
-  } = comments[id];
-
-  return `<li class="film-details__comment">
-    <span class="film-details__comment-emoji">
-      <img src="./images/emoji/${emotion}.png" width="55" height="55" alt="emoji-smile">
-    </span>
-    <div>
-      <p class="film-details__comment-text">${text}</p>
-      <p class="film-details__comment-info">
-        <span class="film-details__comment-author">${author}</span>
-        <span class="film-details__comment-day">
-          ${getFormattedDateTime(date)}
-        </span>
-        <button class="film-details__comment-delete">Delete</button>
-      </p>
-    </div>
-  </li>`;
-};
-
-const isActive = (flag) => flag ? 'checked' : '';
-
-const createFilmDetailsTemplate = (card, comments) => {
-  const {
-    title,
-    poster,
-    description,
-    rating,
-    ageRating,
-    runtime,
-    director,
-    writers,
-    actors,
-    genres,
-    release: {
-      date,
-      country,
-    },
-  } = card.filmInfo;
-
-  const {
-    isWatchlist,
-    isWatched,
-    isFavorite,
-  } = card.userDetails;
-
-  const commentIds = [...card.comments];
-  const commentCount = commentIds.length;
-
-  return `<section class="film-details">
-    <form class="film-details__inner" action="" method="get">
-      <div class="film-details__top-container">
-        <div class="film-details__close">
-          <button class="film-details__close-btn" type="button">close</button>
-        </div>
-        <div class="film-details__info-wrap">
-          <div class="film-details__poster">
-            <img class="film-details__poster-img" src="./images/posters/${poster}" alt="">
-
-            <p class="film-details__age">${ageRating}+</p>
-          </div>
-
-          <div class="film-details__info">
-            <div class="film-details__info-head">
-              <div class="film-details__title-wrap">
-                <h3 class="film-details__title">${title}</h3>
-                <p class="film-details__title-original">${title}</p>
-              </div>
-
-              <div class="film-details__rating">
-                <p class="film-details__total-rating">${rating}</p>
-              </div>
-            </div>
-
-            <table class="film-details__table">
-              <tr class="film-details__row">
-                <td class="film-details__term">Director</td>
-                <td class="film-details__cell">${director}</td>
-              </tr>
-              <tr class="film-details__row">
-                <td class="film-details__term">
-                  ${writers.size === 1 ? 'Writer' : 'Writers'}
-                </td>
-                <td class="film-details__cell">
-                  ${[...writers].join(', ')}
-                </td>
-              </tr>
-              <tr class="film-details__row">
-                <td class="film-details__term">
-                  ${actors.size === 1 ? 'Actor' : 'Actors'}
-                </td>
-                <td class="film-details__cell">
-                  ${[...actors].join(', ')}
-                </td>
-              </tr>
-              <tr class="film-details__row">
-                <td class="film-details__term">Release Date</td>
-                <td class="film-details__cell">
-                  ${getFormattedDate(date)}
-                </td>
-              </tr>
-              <tr class="film-details__row">
-                <td class="film-details__term">Runtime</td>
-                <td class="film-details__cell">
-                  ${getTimeFromMinutes(runtime)}
-                </td>
-              </tr>
-              <tr class="film-details__row">
-                <td class="film-details__term">Country</td>
-                <td class="film-details__cell">${country}</td>
-              </tr>
-              <tr class="film-details__row">
-                <td class="film-details__term">
-                  ${genres.size === 1 ? 'Genre' : 'Genres'}
-                </td>
-                <td class="film-details__cell">
-                  ${[...genres].map((genre) => `<span class="film-details__genre">${genre}</span>`).join(' ')}
-                </td>
-              </tr>
-            </table>
-
-            <p class="film-details__film-description">${description}</p>
-          </div>
-        </div>
-
-        <section class="film-details__controls">
-          <input type="checkbox" class="film-details__control-input visually-hidden" id="watchlist" name="watchlist" ${isActive(isWatchlist)}>
-          <label for="watchlist" class="film-details__control-label film-details__control-label--watchlist">Add to watchlist</label>
-
-          <input type="checkbox" class="film-details__control-input visually-hidden" id="watched" name="watched" ${isActive(isWatched)}>
-          <label for="watched" class="film-details__control-label film-details__control-label--watched">Already watched</label>
-
-          <input type="checkbox" class="film-details__control-input visually-hidden" id="favorite" name="favorite" ${isActive(isFavorite)}>
-          <label for="favorite" class="film-details__control-label film-details__control-label--favorite">Add to favorites</label>
-        </section>
-      </div>
-
-      <div class="film-details__bottom-container">
-        <section class="film-details__comments-wrap">
-          <h3 class="film-details__comments-title">
-            ${commentCount === 1 ? 'Comment' : 'Comments'}
-            <span class="film-details__comments-count">${commentCount}</span>
-          </h3>
-
-          <ul class="film-details__comments-list">
-            ${commentIds.map(getComments(comments)).join('')}
-          </ul>
-
-          <div class="film-details__new-comment">
-            <div class="film-details__add-emoji-label"></div>
-
-            <label class="film-details__comment-label">
-              <textarea class="film-details__comment-input" placeholder="Select reaction below and write comment here" name="comment"></textarea>
-            </label>
-
-            <div class="film-details__emoji-list">
-              <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-smile" value="smile">
-              <label class="film-details__emoji-label" for="emoji-smile">
-                <img src="./images/emoji/smile.png" width="30" height="30" alt="emoji">
-              </label>
-
-              <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-sleeping" value="sleeping">
-              <label class="film-details__emoji-label" for="emoji-sleeping">
-                <img src="./images/emoji/sleeping.png" width="30" height="30" alt="emoji">
-              </label>
-
-              <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-puke" value="puke">
-              <label class="film-details__emoji-label" for="emoji-puke">
-                <img src="./images/emoji/puke.png" width="30" height="30" alt="emoji">
-              </label>
-
-              <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-angry" value="angry">
-              <label class="film-details__emoji-label" for="emoji-angry">
-                <img src="./images/emoji/angry.png" width="30" height="30" alt="emoji">
-              </label>
-            </div>
-          </div>
-        </section>
-      </div>
-    </form>
-  </section>`;
-};
-
-class FilmDetailsView extends AbstractView {
-  constructor(card, comments) {
-    super();
-    this._card = card;
-    this._comments = comments;
-
-    this._closeClickHandler = this._closeClickHandler.bind(this);
-    this._formChangeHandler = this._formChangeHandler.bind(this);
-  }
-
-  getTemplate() {
-    return createFilmDetailsTemplate(this._card, this._comments);
-  }
-
-  _closeClickHandler(evt) {
-    evt.preventDefault();
-
-    this._callback.closeClick(evt);
-  }
-
-  _formChangeHandler(evt) {
-    evt.preventDefault();
-
-    this._callback.formChange(evt);
-  }
-
-  setCloseClickHandler(callback) {
-    this._callback.closeClick = callback;
-
-    this.getElement().querySelector('.film-details__close-btn')
-      .addEventListener('click', this._closeClickHandler);
-  }
-
-  setFormChangeHandler(callback) {
-    this._callback.formChange = callback;
-
-    this.getElement().querySelector('.film-details__inner')
-      .addEventListener('change', this._formChangeHandler);
-  }
-}
-
-export default FilmDetailsView;
